+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Fixed
 - Fix JavaDoc errors.
+- Add new properties `gid.connector.http.sink.error.code` and `gid.connector.http.sink.error.code.exclude`
+  to set HTTP status code that should be interpreted as errors.
 
 ## [0.3.0] - 2022-07-21